requirements.txt
asgiref==3.6.0
attrs==23.1.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
Django~=4.2.5
django-admin-interface==0.26.0
django-ckeditor~=6.7.0
django-colorfield==0.9.0
django-js-asset==2.0.0
django-richtextfield==1.6.1
django-squash==0.0.6
ghasedakpack==0.1.13
idna==3.4
isodate==0.6.1
lxml==4.9.2
mysqlclient==2.1.1
Pillow==9.4.0
platformdirs==3.5.0
python-slugify==8.0.1
pytz==2023.3
requests~=2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
tzdata==2023.2
urllib3==1.26.15
zeep==4.2.1
djangorestframework~=3.14.0